+++ babel-compiler-hot/package.js
@@ -1,7 +1,7 @@
 Package.describe({
   name: "gadicc:babel-compiler-hot",
   // version: '6.5.2-rc.12'    // core version, KEEP UPDATED
-  version: '0.0.15-rc.10',
+  version: '0.0.19-rc.12',
   summary: 'React hotloading, used by gadicc:ecmascript-hot.',
   git: 'https://github.com/gadicc/meteor-react-hotloader',
   documentation: 'README.md'
@@ -11,8 +11,7 @@ Npm.depends({
   //'meteor-babel': '0.8.3',  // core version, KEEP UPDATED
   //'meteor-babel': 'file:///home/dragon/www/npm/gadicc-meteor-babel',
   'meteor-babel': 'https://github.com/gadicc/babel/tarball/deeb76d04b2ab5b9963187914b3242b39858445a',
-  'mongodb': '2.1.4',
-  'json5': '0.5.0'
+  'mongodb': '2.1.4'
 });
 
 Package.onUse(function (api) {
@@ -30,6 +29,7 @@ Package.onUse(function (api) {
   // hot
   api.use('random@1.0.5');
   api.use('underscore@1.0.0');
+  api.use('gadicc:json5@0.5.0-0');
 
   api.export('Babel', 'server');
   api.export('BabelCompiler', 'server');