+++ hot/package.js
@@ -1,6 +1,6 @@
 Package.describe({
   name: 'gadicc:hot',
-  version: '0.0.10',
+  version: '0.0.11',
   summary: 'React hotloading, used by gadicc:ecmascript-hot.',
   git: 'https://github.com/gadicc/meteor-react-hotloader',
   documentation: '../README.md'
@@ -17,7 +17,7 @@ Package.onUse(function(api) {
   api.use('random', 'server');
 
   // this isn't used directly, but is used to pull in the package
-  api.use('gadicc:modules-runtime-hot@0.0.5-rc.4');
+  api.use('gadicc:modules-runtime-hot@0.0.6-rc.12');
   api.use('modules-runtime');
   api.imply('modules-runtime');