+++ modules-runtime-hot/package.js
@@ -1,7 +1,7 @@
 Package.describe({
   name: "gadicc:modules-runtime-hot",
   // version: "0.6.1-rc.12,    // core version, KEEP UPDATED
-  version: "0.0.5-rc.4",
+  version: "0.0.6-rc.12",
   summary: 'Used by gadicc:ecmascript-hot',
   git: "https://github.com/benjamn/install",
   documentation: "README.md"